Refactor Memperbaiki penulisan kode dan menghilangkan folder widgets
