+++ lib/views/dashboard.dart
@@ -3,7 +3,7 @@ import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
 import 'package:parkquest_parkir_gamifikasi/Controllers/AuthenticationController.dart';
 import 'package:parkquest_parkir_gamifikasi/Controllers/LeaderboardController.dart';
-import 'package:parkquest_parkir_gamifikasi/views/widgets/leaderboard_top_three.dart';
+import 'package:parkquest_parkir_gamifikasi/Models/Leaderboard.dart';
 
 class Dashboard extends StatelessWidget {
   Dashboard({super.key});
@@ -107,8 +107,13 @@ class Dashboard extends StatelessWidget {
                       physics: NeverScrollableScrollPhysics(),
                       itemCount: _leaderboardcontroller.datas.value.length,
                       itemBuilder: (context, index) {
-                        return LeaderboardTopThree(
-                          data: _leaderboardcontroller.datas.value[index],
+                        final LeaderboardModel data =
+                            _leaderboardcontroller.datas.value[index];
+                        return Column(
+                          children: [
+                            Text(data.username),
+                            Text(data.rank.rankName.toString()),
+                          ],
                         );
                       },
                     );
