+++ lib/views/inventory/rare.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
 import 'package:parkquest_parkir_gamifikasi/Controllers/InventoryController.dart';
-import 'package:parkquest_parkir_gamifikasi/views/widgets/inventory/rare.dart';
+import 'package:parkquest_parkir_gamifikasi/Models/Avatar/UserAvatar.dart';
 
 class InventoryRare extends StatelessWidget {
   InventoryRare({super.key});
@@ -48,8 +48,15 @@ class InventoryRare extends StatelessWidget {
                       physics: NeverScrollableScrollPhysics(),
                       itemCount: _inventorycontroller.datasRare.value.length,
                       itemBuilder: (context, index) {
-                        return InventoryRareWidget(
-                          data: _inventorycontroller.datasRare.value[index],
+                        final UserAvatarModel data =
+                            _inventorycontroller.datasRare.value[index];
+                        return Column(
+                          children: [
+                            Text(data.rarityName),
+                            Text(data.avatarName),
+                            Text(data.avatarImage),
+                            Text(data.avatarDescription),
+                          ],
                         );
                       },
                     );
