+++ lib/views/leaderboard.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
 import 'package:parkquest_parkir_gamifikasi/Controllers/LeaderboardController.dart';
-import 'package:parkquest_parkir_gamifikasi/views/widgets/leaderboard/leaderboard.dart';
+import 'package:parkquest_parkir_gamifikasi/Models/Leaderboard.dart';
 
 class Leaderboard extends StatelessWidget {
   Leaderboard({super.key});
@@ -36,8 +36,13 @@ class Leaderboard extends StatelessWidget {
                       physics: NeverScrollableScrollPhysics(),
                       itemCount: _leaderboardcontroller.datas.value.length,
                       itemBuilder: (context, index) {
-                        return LeaderboardWidget(
-                          data: _leaderboardcontroller.datas.value[index],
+                        final LeaderboardModel data =
+                            _leaderboardcontroller.datas.value[index];
+                        return Column(
+                          children: [
+                            Text(data.username),
+                            Text(data.rank.rankName.toString()),
+                          ],
                         );
                       },
                     );
