+++ lib/views/shop/basic.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
 import 'package:parkquest_parkir_gamifikasi/Controllers/ShopController.dart';
-import 'package:parkquest_parkir_gamifikasi/views/widgets/shop/basic.dart';
+import 'package:parkquest_parkir_gamifikasi/Models/Avatar/Avatar.dart';
 
 class ShopBasic extends StatelessWidget {
   ShopBasic({super.key});
@@ -47,8 +47,16 @@ class ShopBasic extends StatelessWidget {
                       physics: NeverScrollableScrollPhysics(),
                       itemCount: _shopcontroller.datasBasic.value.length,
                       itemBuilder: (context, index) {
-                        return ShopBasicWidget(
-                          data: _shopcontroller.datasBasic.value[index],
+                        final AvatarModel data =
+                            _shopcontroller.datasBasic.value[index];
+
+                        return Column(
+                          children: [
+                            Text(data.rarityName),
+                            Text(data.avatarName),
+                            Text(data.avatarImage),
+                            Text(data.avatarDescription),
+                          ],
                         );
                       },
                     );
